some general improvements (#99)
closes #85

+++ src/main/java/ch/hslu/cas/msed/blobfish/player/bot/minimax/MoveNode.java
@@ -1,3 +1,3 @@
 package ch.hslu.cas.msed.blobfish.player.bot.minimax;
 
-public record MoveNode(double eval, MoveHistoryNode history) { }
+public record MoveNode(int eval, MoveHistoryNode history) { }
